Initial pass at setting up Travis CI and coveralls
Tox is known to behae badly with anaconda. Where tox would normally be used
I've substituted a bash script. To simulate what the TavisCI agent will run use
./provision/travis/run-tests.sh

It would be good to find a tox alternative. (Or failing that we could use a different method
of installing dependencies e.g. docker or using the apt-get version of FENICS...)